reset (user_register) register and (token/login/get_self_user_info)login flow

+++ app/src/main/java/com/ti4n/freechat/model/response/UserToken.kt
@@ -1,7 +1,7 @@
 package com.ti4n.freechat.model.response
 
 data class UserToken(
-    val expiredTime: Int,
+    val expiredTime: Long,
     val token: String,
     val userID: String
 )
